requirements.txt
requests==2.23.0
pycryptodome==3.11.0
